Changes after review
	Renamed h && cc file names in telemetry_monitor.
	Corrected header-guards.
	Corrected commented code in gmock.
	Created TODO and tasks for them.
