+++ src/components/include/test/protocol_handler/mock_protocol_handler.h
@@ -63,4 +63,4 @@ class MockProtocolHandler : public ::protocol_handler::ProtocolHandler {
 }  // namespace components
 }  // namespace test
 
-#endif  // SRC_COMPONENTS_INCLUDE_TEST_PROTOCOL_HANDLER_MOCK_PROTOCOL_HANDLER_SETTINGS_H_
+#endif  // SRC_COMPONENTS_INCLUDE_TEST_PROTOCOL_HANDLER_MOCK_PROTOCOL_HANDLER_H_
