+++ src/components/security_manager/include/security_manager/crypto_manager_settings_impl.h
@@ -19,23 +19,25 @@ class CryptoManagerSettingsImpl : public CryptoManagerSettings {
   }
   Protocol security_manager_protocol_name() const OVERRIDE {
     CREATE_LOGGERPTR_LOCAL(logger_, "SecurityManager")
-    security_manager::Protocol protocol;
+
     const std::string& protocol_str = profile_.security_manager_protocol_name();
-    if (profile_.security_manager_protocol_name() == "TLSv1.0") {
-      protocol = security_manager::TLSv1;
-    } else if (protocol_str == "TLSv1.1") {
-      protocol = security_manager::TLSv1_1;
-    } else if (protocol_str == "TLSv1.2") {
-      protocol = security_manager::TLSv1_2;
-    } else if (protocol_str == "SSLv3") {
-      protocol = security_manager::SSLv3;
-    } else {
-      LOG4CXX_ERROR(
+    if (protocol_str == "TLSv1.0") {
+      return security_manager::TLSv1;
+    }
+    if (protocol_str == "TLSv1.1") {
+      return security_manager::TLSv1_1;
+    }
+    if (protocol_str == "TLSv1.2") {
+      return security_manager::TLSv1_2;
+    }
+    if (protocol_str == "SSLv3") {
+      return security_manager::SSLv3;
+    }
+    LOG4CXX_ERROR(
           logger_,
           "Unknown protocol: " << profile::Profile::instance()
                                       ->security_manager_protocol_name());
-    }
-    return protocol;
+    return static_cast<security_manager::Protocol>(-1);
   }
   bool verify_peer() const OVERRIDE {
     return profile_.verify_peer();
