+++ src/components/telemetry_monitor/src/application_manager_observer.cc
@@ -32,7 +32,7 @@
 #include "telemetry_monitor/application_manager_observer.h"
 #include "utils/shared_ptr.h"
 #include "telemetry_monitor/telemetry_monitor.h"
-#include "telemetry_monitor/application_manager_metric.h"
+#include "telemetry_monitor/application_manager_metric_wrapper.h"
 
 namespace telemetry_monitor {
