+++ src/components/telemetry_monitor/src/transport_manager_observer.cc
@@ -33,7 +33,7 @@
 
 #include <time.h>
 
-#include "telemetry_monitor/transport_manager_metric.h"
+#include "telemetry_monitor/transport_manager_metric_wrapper.h"
 #include "telemetry_monitor/telemetry_monitor.h"
 
 namespace telemetry_monitor {
