+++ src/components/telemetry_monitor/test/application_manager_metric_test.cc
@@ -34,7 +34,7 @@
 #include "telemetry_monitor/json_keys.h"
 #include "utils/resource_usage.h"
 #include "application_manager/smart_object_keys.h"
-#include "telemetry_monitor/application_manager_metric.h"
+#include "telemetry_monitor/application_manager_metric_wrapper.h"
 
 namespace test {
 namespace components {
