+++ src/components/telemetry_monitor/test/application_manager_observer_test.cc
@@ -33,7 +33,7 @@
 #include "gtest/gtest.h"
 #include "telemetry_monitor/mock_telemetry_monitor.h"
 #include "application_manager/telemetry_observer.h"
-#include "telemetry_monitor/application_manager_metric.h"
+#include "telemetry_monitor/application_manager_metric_wrapper.h"
 #include "telemetry_monitor/application_manager_observer.h"
 #include "utils/shared_ptr.h"
 #include "telemetry_monitor/telemetry_monitor.h"
