+++ src/components/telemetry_monitor/test/transport_manager_metric_test.cc
@@ -34,7 +34,7 @@
 #include "telemetry_monitor/json_keys.h"
 #include "json/json.h"
 #include "utils/resource_usage.h"
-#include "telemetry_monitor/transport_manager_metric.h"
+#include "telemetry_monitor/transport_manager_metric_wrapper.h"
 #include "protocol_handler/telemetry_observer.h"
 
 namespace test {
