+++ src/components/telemetry_monitor/test/transport_manager_observer_test.cc
@@ -33,7 +33,7 @@
 #include "gtest/gtest.h"
 #include "telemetry_monitor/mock_telemetry_monitor.h"
 #include "protocol_handler/telemetry_observer.h"
-#include "telemetry_monitor/transport_manager_metric.h"
+#include "telemetry_monitor/transport_manager_metric_wrapper.h"
 #include "telemetry_monitor/transport_manager_observer.h"
 
 namespace test {
